+++ source/RTSCamera/src/View/FlyCameraMissionView.cs
@@ -497,7 +497,7 @@ namespace RTSCamera.View
                 if (RTSCameraGameKeyCategory.GetKey(GameKeyEnum.CameraMoveDown).IsKeyDown())
                     --keyInputVertical;
 
-                if (MissionScreen.MouseVisible && !MissionScreen.SceneLayer.Input.IsKeyDown(InputKey.RightMouseButton))
+                if (MissionScreen.MouseVisible && !MissionScreen.SceneLayer.Input.IsKeyDown(InputKey.RightMouseButton) && !RTSCameraGameKeyCategory.GetKey(GameKeyEnum.CameraMoveByMouse).IsKeyDown())
                 {
                     if (Mission.Mode != MissionMode.Conversation)
                     {
@@ -556,15 +556,6 @@ namespace RTSCamera.View
 
             if (_classicMode)
             {
-                cameraFrame.origin += _cameraSpeed.x * cameraFrame.rotation.s.AsVec2.ToVec3().NormalizedCopy() * dt;
-                ref Vec3 local = ref cameraFrame.origin;
-                Vec3 vec3_2 = local;
-                float y = _cameraSpeed.y;
-                Vec3 vec3_1 = cameraFrame.rotation.u.AsVec2.ToVec3();
-                Vec3 vec3_3 = vec3_1.NormalizedCopy();
-                Vec3 vec3_4 = y * vec3_3 * dt;
-                local = vec3_2 - vec3_4;
-                cameraFrame.origin.z += _cameraSpeed.z * dt;
                 float mouseScroll = Input.GetDeltaMouseScroll();
                 float controllerHeightInput = 0;
                 if (MissionScreen.SceneLayer.Input.IsKeyDown(InputKey.ControllerLTrigger))
@@ -574,6 +565,7 @@ namespace RTSCamera.View
                         controllerHeightInput = 0.0f;
                 }
                 var diffRatio = MathF.Min(1f, dt * 5f);
+                Vec3 cameraMovement = Vec3.Zero;
                 if (FocusedFormation != null)
                 {
                     _lookingDistanceToAdd -= (mouseScroll / 200.0f + controllerHeightInput) * cameraBasicSpeed * VerticalMovementSpeedFactor;
@@ -605,22 +597,28 @@ namespace RTSCamera.View
                     if (!isSpeedKeyDown)
                         _cameraHeightToAdd -= (mouseScroll / 2000.0f + controllerHeightInput / 100f) * verticalLimit;
                     // hold middle button and move mouse vertically to adjust height
-                    if (MissionScreen.SceneLayer.Input.IsHotKeyDown("DeploymentCameraIsActive"))
+                    if (RTSCameraGameKeyCategory.GetKey(GameKeyEnum.CameraMoveByMouse).IsKeyDown())
                     {
                         if (_levelToEdge == false)
                         {
                             _levelToEdge = true;
-                            ScreenManager.FirstHitLayer.InputRestrictions.SetMouseVisibility(true);
+                            //_showControlHintLayer.InputRestrictions.SetMouseVisibility(true);
                         }
-                        _cameraHeightToAdd += 0.5f * TaleWorlds.InputSystem.Input.MouseMoveY;
+                        //_cameraHeightToAdd += 0.5f * TaleWorlds.InputSystem.Input.MouseMoveY;
+                        cameraMovement.y += TaleWorlds.InputSystem.Input.MouseMoveY * 0.1f;
+                        cameraMovement.x -= TaleWorlds.InputSystem.Input.MouseMoveX * 0.1f;
                     }
                     else if (_levelToEdge)
                     {
                         _levelToEdge = false;
-                        ScreenManager.FirstHitLayer.InputRestrictions.SetMouseVisibility(false);
+                        //_showControlHintLayer.InputRestrictions.SetMouseVisibility(false);
                     }
                     _cameraHeightToAdd = MathF.Clamp(_cameraHeightToAdd, -verticalLimit, verticalLimit);
                 }
+                cameraFrame.origin += (_cameraSpeed.x * dt + cameraMovement.x) * cameraFrame.rotation.s.AsVec2.ToVec3().NormalizedCopy();
+                Vec3 vec3_1 = (_cameraSpeed.y * dt + cameraMovement.y) * cameraFrame.rotation.u.AsVec2.ToVec3().NormalizedCopy();
+                cameraFrame.origin -= vec3_1;
+                cameraFrame.origin.z += _cameraSpeed.z * dt;
                 if (MathF.Abs(_cameraHeightToAdd) > 1.0 / 1000.0)
                 {
                     cameraFrame.origin.z += _cameraHeightToAdd * diffRatio;
@@ -822,7 +820,7 @@ namespace RTSCamera.View
                         inputYRaw += (float)-num3 * MissionScreen.SceneLayer.Input.GetGameKeyAxis("CameraAxisY");
                     }
                 }
-                else if (!MissionScreen.MouseVisible)
+                else if (!MissionScreen.MouseVisible && !RTSCameraGameKeyCategory.GetKey(GameKeyEnum.CameraMoveByMouse).IsKeyDown())
                 {
                     inputXRaw = MissionScreen.SceneLayer.Input.GetMouseMoveX();
                     inputYRaw = MissionScreen.SceneLayer.Input.GetMouseMoveY();